Instagram-Website/requirements.txt
Django>=3.2,<4.0
djangorestframework>=3.12
django-cors-headers>=3.7
psycopg2-binary>=2.8  # Use this if connecting to PostgreSQL
